Unit Tests movies usecases

+++ domain/src/main/java/com/ludovic/vimont/domain/repositories/MovieRepository.kt
@@ -7,7 +7,7 @@ import com.ludovic.vimont.domain.entities.Movie
 interface MovieRepository {
     fun get(id: Int): StateData<Movie>
 
-    fun discover(voteAverage: Float, genres: List<Genre> = ArrayList()): StateData<Movie>
+    fun discover(voteAverage: Double, genres: List<Genre> = ArrayList()): StateData<Movie>
 
     fun search(query: String): StateData<List<Movie>>
 }
